% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotScatterPie.R
\name{plotScatterPie}
\alias{plotScatterPie}
\title{plotScatterPie}
\usage{
plotScatterPie(data, xmeanLine = TRUE, ymeanLine = TRUE, labelLine = NULL)
}
\arguments{
\item{data}{a dataframe with information like x, y, r, etc. Refer to \code{GDPmix}}

\item{xmeanLine}{if plot x mean line}

\item{ymeanLine}{if plot y mean line}

\item{labelLine}{how far is label to pie chart, can be left with default value.}
}
\description{
plot scatter pie chart for multidimension analysis, such as waternomics. This plot can
provide information about water use/wastewater of each provinces and GDP mix of each provinces,
see examples.
}
